+++ Runtime/Levels/LevelData.cs
@@ -6,6 +6,7 @@ namespace Pixygon.Micro
     [CreateAssetMenu(menuName = "Pixygon/Micro/New LevelData")]
     public class LevelData : ScriptableObject {
         public string _levelName;
+        public string _levelId;
         public AssetReference _levelRef;
         public AssetReference _bgmRef;
         public AssetReference _postProcessingProfileRef;